chore: move the hook to consign directory
